require explicit dependency on editor for typings
