+++ ui/analyse/tsconfig.json
@@ -8,6 +8,7 @@
     { "path": "../chat/tsconfig.json" },
     { "path": "../chess/tsconfig.json" },
     { "path": "../common/tsconfig.json" },
+    { "path": "../editor/tsconfig.json" },
     { "path": "../game/tsconfig.json" },
     { "path": "../keyboardMove/tsconfig.json" },
     { "path": "../nvui/tsconfig.json" },
