+++ ui/editor/package.json
@@ -11,6 +11,7 @@
   ],
   "author": "Thibault Duplessis",
   "license": "AGPL-3.0-or-later",
+  "types": "dist/editor.d.ts",
   "dependencies": {
     "chessops": "^0.14.1",
     "common": "workspace:*",
