+++ ui/editor/src/editor.ts
@@ -1,11 +1,14 @@
 import EditorCtrl from './ctrl';
+import { LichessEditor, Config } from './interfaces';
 import menuHover from 'common/menuHover';
 import view from './view';
 import { init, attributesModule, eventListenersModule, classModule, propsModule } from 'snabbdom';
 
 const patch = init([classModule, attributesModule, propsModule, eventListenersModule]);
 
-export function initModule(config: Editor.Config): LichessEditor {
+export type { LichessEditor } from './interfaces';
+
+export function initModule(config: Config): LichessEditor {
   const ctrl = new EditorCtrl(config, redraw);
 
   const el = config.el || document.getElementById('board-editor')!;
