+++ ui/editor/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../tsconfig.base.json",
-  "compilerOptions": { "noEmit": true },
+  "compilerOptions": { "emitDeclarationOnly": true },
   "references": [{ "path": "../common/tsconfig.json" }]
 }
